// dx/frontend/src/api/model/mediaItemOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * dx API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaItemOut {
  /** @maxLength 100 */
  content_type: string;
  created: string;
  id: string;
  /** @maxLength 5 */
  kind: string;
  /** @maxLength 255 */
  name: string;
  size: number;
  url: string;
}